+++ src/main/java/vectorwing/farmersdelight/data/BlockStates.java
@@ -149,6 +149,7 @@ public class BlockStates extends BlockStateProvider
 		this.cabinetBlock(ModBlocks.JUNGLE_CABINET.get(), "jungle");
 		this.cabinetBlock(ModBlocks.ACACIA_CABINET.get(), "acacia");
 		this.cabinetBlock(ModBlocks.DARK_OAK_CABINET.get(), "dark_oak");
+		this.cabinetBlock(ModBlocks.MANGROVE_CABINET.get(), "mangrove");
 		this.cabinetBlock(ModBlocks.CRIMSON_CABINET.get(), "crimson");
 		this.cabinetBlock(ModBlocks.WARPED_CABINET.get(), "warped");